Trabajado en clase

+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/JugadorDeCampo.java
@@ -9,7 +9,7 @@ public class JugadorDeCampo extends Jugador{
 
 	private Integer golesConvertidos;
 	
-	public JugadorDeCampo(Long id, String nombre, Integer edad, Date fechaDeNacimiento, Double peso, Double altura, Integer valoracion, Integer golesConvertidos) {
+	public JugadorDeCampo(Long id, String nombre, Integer edad, Date fechaDeNacimiento, Double peso, Double altura, Double valoracion, Integer golesConvertidos) {
 		super(id, nombre, edad, fechaDeNacimiento, peso, altura, valoracion);
 		
 		this.golesConvertidos = golesConvertidos;
